remove local state from editor and set defaults when adding a node
